float speeds instead of ints

+++ HW3/src/components/Road.java
@@ -186,7 +186,7 @@ public class Road implements RouteParts {
 	@Override
 	public double calcEstimatedTime(Object obj) {
 		Vehicle v = (Vehicle) obj;
-		int speed = Math.min(maxSpeed, v.getVehicleType().getAverageSpeed());
+		float speed = Math.min(maxSpeed, v.getVehicleType().getAverageSpeed());
 		return (int) length / speed;
 	}
